backend/app/requirements-dev.txt
fastapi
uvicorn
sqlalchemy[asyncio]
alembic
psycopg2-binary
python-jose
asyncpg
python-dotenv
pytest-asyncio
httpx
requests
itsdangerous
python-multipart
